Tweaking

+++ src/it/unimi/dsi/sux4j/bits/Fast.java
@@ -100,7 +100,7 @@ public final class Fast {
 	
 	public static void main( final String a[] ) {
 		final long n = Long.parseLong( a[ 0 ] );
-		final long incr = Long.MAX_VALUE / ( n / 2 );
+		final long incr = Integer.MAX_VALUE / ( n / 2 );
 		
 		long start, elapsed;
 		
@@ -108,7 +108,7 @@ public final class Fast {
 			System.out.print( "Broadword: " );
 			
 			start = System.currentTimeMillis();
-			for( long i = n, v = 0; i-- != 0; ) mostSignificantBit( v += incr );
+			for( int i = (int)n, v = 0; i-- != 0; ) mostSignificantBit( v += incr );
 			elapsed = System.currentTimeMillis() - start;
 
 			System.out.println( "elapsed " + elapsed + ", " + ( 1000000.0 * elapsed / n ) + " ns/call" );
@@ -116,7 +116,7 @@ public final class Fast {
 			System.out.print( "Dichotomous: " );
 			
 			start = System.currentTimeMillis();
-			for( long i = n, v = 0; i-- != 0; ) it.unimi.dsi.mg4j.util.Fast.mostSignificantBit( v += incr );
+			for( int i = (int)n, v = 0; i-- != 0; ) it.unimi.dsi.mg4j.util.Fast.mostSignificantBit( v += incr );
 			elapsed = System.currentTimeMillis() - start;
 
 			System.out.println( "elapsed " + elapsed + ", " + ( 1000000.0 * elapsed / n ) + " ns/call" );
@@ -124,11 +124,12 @@ public final class Fast {
 			System.out.print( "java.lang: " );
 			
 			start = System.currentTimeMillis();
-			for( long i = n, v = 0; i-- != 0; ) Long.numberOfLeadingZeros( v += incr );
+			for( int i = (int)n, v = 0; i-- != 0; ) Long.numberOfLeadingZeros( v += incr );
 			elapsed = System.currentTimeMillis() - start;
 
 			System.out.println( "elapsed " + elapsed + ", " + ( 1000000.0 * elapsed / n ) + " ns/call" );
 		}
+
 	}
 	
 }
